#227 Extract the Via-join engine out of QueryPlanner
Move executeViaPlan, perParentLoop, and the six hashJoin* methods into an internal
ViaJoinExecutor, leaving QueryPlanner focused on planning (strategy selection, index
candidates, range slicing, ordering/pagination). The ViaStrategy enum stays in QueryPlanner
as part of its test-visible contract.
